% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuss_COVARIATE_mcmc.R
\name{check_convergence}
\alias{check_convergence}
\title{Check MCMC convergence}
\usage{
check_convergence(object, chains = NULL)
}
\arguments{
\item{object}{A \code{fuss_COVARIATE_mcmc} object}

\item{chains}{Optional list of multiple chains for Gelman-Rubin}
}
\value{
Convergence diagnostics
}
\description{
Check MCMC convergence
}
