% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{check_PACKS}
\alias{check_PACKS}
\title{Check if required packages are available}
\usage{
check_PACKS()
}
\value{
Invisibly returns TRUE if all packages are available
}
\description{
Check if required packages are available
}
