% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{compare_parallel_methods}
\alias{compare_parallel_methods}
\title{Compare both MCMC methods in parallel}
\usage{
compare_parallel_methods(data, varCLASS, varY, varID, n_workers = 3, ...)
}
\arguments{
\item{data}{Input data frame}

\item{varCLASS}{Category variable name}

\item{varY}{Value variable name}

\item{varID}{ID variable name}

\item{n_workers}{Number of parallel workers}

\item{...}{Additional parameters}
}
\value{
List with both results
}
\description{
Compare both MCMC methods in parallel
}
