% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuss_COVARIATE_mcmc.R
\name{effective_sample_size}
\alias{effective_sample_size}
\title{Compute Effective Sample Size (ESS) for MCMC chains}
\usage{
effective_sample_size(x)
}
\arguments{
\item{x}{Numeric vector or matrix of MCMC samples}
}
\value{
Effective sample size
}
\description{
Compute Effective Sample Size (ESS) for MCMC chains
}
