% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{fuss_DIRICHLET}
\alias{fuss_DIRICHLET}
\title{Quick wrapper for Dirichlet parallel processing}
\usage{
fuss_DIRICHLET(...)
}
\arguments{
\item{...}{Same parameters as fuss_PARALLEL}
}
\value{
Same output format
}
\description{
Quick wrapper for Dirichlet parallel processing
}
