% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{fuss_PARALLEL_dirichlet}
\alias{fuss_PARALLEL_dirichlet}
\title{Parallel Dirichlet wrapper - works with fuss_PARALLEL}
\usage{
fuss_PARALLEL_dirichlet(..., mcmc_method = "dirichlet", dirichlet_alpha = 2)
}
\arguments{
\item{...}{All parameters from fuss_PARALLEL}

\item{mcmc_method}{"dirichlet" or "metropolis" (default)}

\item{dirichlet_alpha}{Dirichlet concentration parameter}
}
\value{
Same as fuss_PARALLEL
}
\description{
Parallel Dirichlet wrapper - works with fuss_PARALLEL
}
