% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_MODES}
\alias{get_MODES}
\title{Obtain modes using various bandwidth selection methods}
\usage{
get_MODES(y)
}
\arguments{
\item{y}{vector, input data of a distribution}
}
\value{
data frame with estimated modes, methods
}
\description{
Obtain modes using various bandwidth selection methods
}
