% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_MODES_enhanced}
\alias{get_MODES_enhanced}
\title{Enhanced mode detection}
\usage{
get_MODES_enhanced(y, adjust = 1, threshold = 1)
}
\arguments{
\item{y}{Numeric vector}

\item{adjust}{Bandwidth adjustment factor}

\item{threshold}{Relative threshold for significant peaks}
}
\value{
List with mode estimates from multiple methods
}
\description{
Enhanced mode detection
}
