% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_NGRP}
\alias{get_NGRP}
\title{Obtain the number of groups based on bandwidth selection}
\usage{
get_NGRP(y)
}
\arguments{
\item{y}{vector, input data of a distribution}
}
\value{
data frame with methods and number of groups
}
\description{
Obtain the number of groups based on bandwidth selection
}
