% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_NGRP_2d}
\alias{get_NGRP_2d}
\title{2D clustering for detecting number of groups using y-values and density}
\usage{
get_NGRP_2d(y, method = "hdbscan", minPts = 5, eps = 0.2)
}
\arguments{
\item{y}{vector, input data of a distribution}

\item{method}{clustering method ("hdbscan", "dbscan", "optics", "kmeans_2d")}

\item{minPts}{minimum points for density-based clustering (default: 5)}

\item{eps}{epsilon parameter for DBSCAN/OPTICS (default: 0.2)}
}
\value{
integer number of groups detected
}
\description{
2D clustering for detecting number of groups using y-values and density
}
