% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{group_MODES}
\alias{group_MODES}
\title{Group/merge modes if they are within a given range}
\usage{
group_MODES(df, within = 0.03)
}
\arguments{
\item{df}{data frame containing samples of a distribution in its 'Est_Mode' variable}

\item{within}{numeric, range for grouping modes}
}
\value{
data frame with grouped modes
}
\description{
Group/merge modes if they are within a given range
}
