% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{init_gmm_fast}
\alias{init_gmm_fast}
\title{Fast Gaussian Mixture Model initialization using k-means++}
\usage{
init_gmm_fast(y, k = NULL, max_k = 5, n_init = 10)
}
\arguments{
\item{y}{vector, input data}

\item{k}{number of components (if NULL, uses adaptive selection)}

\item{max_k}{maximum components to consider (default: 5)}

\item{n_init}{number of k-means initializations (default: 10)}
}
\value{
list with means, sds, and weights
}
\description{
Fast Gaussian Mixture Model initialization using k-means++
}
\keyword{internal}
