\name{MultiModalR-package}
\alias{MultiModalR-package}
\alias{MultiModalR}
\docType{package}
\title{
\packageTitle{MultiModalR}
}
\description{
\packageDescription{MultiModalR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MultiModalR}
\packageIndices{MultiModalR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{MultiModalR}

Maintainer: \packageMaintainer{MultiModalR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
