% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{print.MultiModalR}
\alias{print.MultiModalR}
\title{MultiModalR Result Object}
\usage{
\method{print}{MultiModalR}(x, ...)
}
\arguments{
\item{x}{A MultiModalR result object}

\item{...}{Additional arguments}
}
\description{
MultiModalR Result Object
}
