% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_MODES}
\alias{get_MODES}
\title{Extract mode statistics from a mode forest}
\usage{
get_MODES(y, adjust = 1)
}
\arguments{
\item{y}{vector, input data of a distribution}

\item{adjust}{adjustment factor for bandwidth (default: 1.0)}
}
\value{
data frame with estimated modes
}
\description{
Extract mode statistics from a mode forest
}
