% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiModalR.R
\name{get_MODES}
\alias{get_MODES}
\title{Extract mode statistics from a mode forest}
\usage{
get_MODES(y, nmod)
}
\arguments{
\item{y}{vector, input data of a distribution}

\item{nmod}{numeric, count/number of subpopulations/subgroups}
}
\value{
data frame with estimated modes
}
\description{
Extract mode statistics from a mode forest
}
