% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{multimodal_dummy}
\alias{multimodal_dummy}
\title{Multimodal Dummy Dataset}
\format{
A data frame with 675 rows and 4 columns
}
\usage{
data(multimodal_dummy)
}
\description{
A simulated dataset containing 9 categories each with 3 distinct 
subpopulations following truncated normal distributions. Ideal for testing 
multimodal mixture modeling.
}
\keyword{datasets}
